Refactor fixture
